git-svn-id: svn://192.168.1.10/public/CoffeeMud@13737 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Items/BasicTech/StdElecItem.java
@@ -43,7 +43,7 @@ public class StdElecItem extends StdItem implements Electronics
 		return "StdElecItem";
 	}
 	
-	protected int			maxRechargePer	= Integer.MAX_VALUE;
+	protected long			maxRechargePer	= Long.MAX_VALUE;
 	protected long			powerCapacity	= 100;
 	protected long			power			= 100;
 	protected boolean		activated		= false;
